Added new features to fights plus fixed somethings
Added it so that you can be rewarded an item, coins, or both at the end of a battle. You will also now receive exp at the end of battles. And in addition, I fixed the problem with the very bad way of detecting enter being pressed; by not detecting enter and just detecting anything in else. Now, I made it so that it checks if the variable is empty, if so then it considers it as you hitting enter. This also fixes a bug where it just doesn't detect anything.