Alpha (#2)
* main branches file replaced with beta file

* Cleaned up code

* Fixed traveling_merchant

Fixed the selling and buying from the merchant

* Added class "Shop"

And integrated it into traveling_merchant

* Began working on back_alley and fixed wasted effect on drinking

* Added shop_list

Combined all the separate dictionaries, into one, big, dictionary that holds all the items.

Made Class Shop be able either pick the items you want to sell, or, if not, it will randomize from the dictionary items.

Temporarily removed wasted effects on other drinks, other than Dragon's Blood.

* Cleaned up some code and attempted to fix the shop

Also cleaned up some of the dictionaries.

* Fixed some code fixing and finished the back_alley scene

* Removed Durability

Removed the durability feature from weapons. Mainly because of the difficulty working with it, but also because it seemed pretty useless.

* Cleaned-up code

No bug fixes or anything, literally just cleaned-up the code.

* Added new features to fights plus fixed somethings

Added it so that you can be rewarded an item, coins, or both at the end of a battle. You will also now receive exp at the end of battles. And in addition, I fixed the problem with the very bad way of detecting enter being pressed; by not detecting enter and just detecting anything in else. Now, I made it so that it checks if the variable is empty, if so then it considers it as you hitting enter. This also fixes a bug where it just doesn't detect anything.

* Hot fix and feature finished

- Finished alcohol system, can now get drunk and appear in alley, subtracts money from coin, actually added wine into drinks subdictionary.
- Fixed healing with health_potions
- Changed input var == "" into not input var, which is cleaner and fixed a bug.
- Fixed grammer

* Delete LICENSE

* Create LICENSE

* Create README.md

* Update README.md

Got rid of link to Alpha branch

* pre-1.0.1-alpha

- Fixed `input var == ""` to `not input var`
- Tweaked wine's `drunk_effective` variable from `.45` to `.25`.
- Fixed time from `.25` to `.5`.
- Changed `thug_fight` in `kidnappers` function to `kidnappers_fight` to differentiate over `thug_fight` in `dark_alley` function.
- Balanced `thug_fight` and `kidnappers_fight`'s exp amounts.
- Grammar corrections